Create "blog" component and order management page. Adjust role handling.
Blog editor is a WYSIWYG editor based on draft.js. Ended up rendering
submitted posts in the same editor window but readonly and with controls
hidden to preserve styling. Need to implement editing functionality.
Probably a "display" toggle and the ability to schedule would also help.
Pictures are limited to ones already hosted but an upload handler
could be used to drop images on the server and respond with a link
to embed.

Order system currently consists of viewing orders, with a reject/confirm
option on the producer side and a cancel option on the retailer side.
Still need to implement order editing but unsure of how to proceed.
Also probably worth considering the issue of race conditions with
respects to order edits.