Added edit functionality to blog component. Created stock categories.
Blog editing is now possible along with a display checkbox to hide
posts if desired. Need to fix the pagination controls to get the
correct post quantity after filtering. Stock categories are now
implemented along with the ability to display them in a custom
order on the profile page. Would be good to limit the amount of
tables to the amount of categories. Might also move the 'add' control
to the bottom of the table.